Will display the file data of the various duplicated photos when a group of duplicated photos is clicked.
Utils/localSys.py methods added to consistently use the same slashes (forward / back) when in Windows

controller.py new methods now assume the responsibility for preparing the data that will later be displayed by view.py

fileMetadata.py new method now shows the relative path of the file found in reference to the folder that was searched.  For this it must also store the folder that was searched.

model.py altered so that it will store in an instance of the FileMetadata the folder that was searched.

view.py now uses a treeview to show the path to the files.  As a result, we have refactored the name.  This change also implements new methods that displays these details in the treeview.